pyproject.toml
[tool.poetry]
name = "multiapptsimulationsanalysis"
version = "0.1.0"
description = ""
authors = ["Benjamin Ravenscroft <benjamin.ravenscroft@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
pyarrow = "^18.0.0"
typer = "^0.13.0"
python-dotenv = "^1.0.1"
tqdm = "^4.67.0"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
